pyproject.toml
[project]
name = "penfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.1",
    "django>=5.2.3",
    "django-ninja-jwt>=5.3.7",
    "django-ninja[jwt]>=1.4.3",
    "python-decouple>=3.8",
]
